% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_notie_alert.R
\name{add_notie_alert}
\alias{add_notie_alert}
\title{Implemente la fonction alert de la librairie JS notie}
\usage{
add_notie_alert(
  type = "success",
  text = "",
  stay = FALSE,
  time = 3,
  position = "top",
  session
)
}
\arguments{
\item{type}{type d'alerte : string ('success', 'warning', 'error', 'info', 'neutral')}

\item{text}{texte affiche dans l'alerte}

\item{stay}{la notif reste elle affichee sur l'ecran (boolean)}

\item{time}{duree d'affichage de la notif}

\item{position}{position de la notif : string ('top', 'bottom')}

\item{session}{argument session shiny}
}
\value{
une alerte notie
}
\description{
Implemente la fonction alert de la librairie JS notie
}
\details{
https://github.com/jaredreich/notie

https://codingshiksha.com/javascript/notie-js-javascript-alert-toast-notification-library-full-tutorial-with-examples/
}
\examples{
\dontrun{library(shiny)
library(shinyYM)

ui <- fluidPage(
 add_notie_deps(),
 actionButton(inputId = "b1", "alert success"),
 actionButton(inputId = "b2", "alert warning"),
 actionButton(inputId = "b3", "alert error"),
 actionButton(inputId = "b4", "alert info"),
 actionButton(inputId = "b5", "alert neutral")
)

server <- function(input, output, session) {

 observeEvent(input$b1,
               add_notie_alert(type = "success", text = "success",
                stay = FALSE, time = 3, position = "top", session)
  )

  observeEvent(input$b2,
               add_notie_alert(type = "warning", text = "warning",
                stay = TRUE, time = 3, position = "bottom", session)
  )

  observeEvent(input$b3,
               add_notie_alert(type = "error", text = "error",
                stay = TRUE, time = 1, position = "bottom", session)
  )

  observeEvent(input$b4,
               add_notie_alert(type = "info", text = "info",
                stay = FALSE, time = 5, position = "top", session)
  )

  observeEvent(input$b5,
               add_notie_alert(type = "neutral", text = "neutral",
                stay = FALSE, time = 3, position = "bottom", session)
  )

}

# Run the application
shinyApp(ui = ui, server = server)
}
}
