% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_notie_date_dialog.R
\name{add_notie_date_dialog}
\alias{add_notie_date_dialog}
\title{Implemente la fonction date de la librairie JS notie}
\usage{
add_notie_date_dialog(
  inputId = "date",
  submitText = "Valider",
  cancelText = "Annuler",
  position = "top",
  session
)
}
\arguments{
\item{inputId}{id qui va contenir la reponse de l'appel à la notif (TRUE/FALSE)}

\item{submitText}{texte de la case de validation de la notif}

\item{cancelText}{texte de la case de refus de la notif}

\item{position}{position de la notif : string ('top', 'bottom')}

\item{session}{argument session shiny}
}
\value{
une alerte notie
}
\description{
Implemente la fonction date de la librairie JS notie
}
\details{
https://github.com/jaredreich/notie

https://codingshiksha.com/javascript/notie-js-javascript-alert-toast-notification-library-full-tutorial-with-examples/
}
\examples{
\dontrun{library(shiny)
library(shinyYM)

ui <- fluidPage(
  add_notie_deps(),
  actionButton(inputId = "b1", "confirm"),
  verbatimTextOutput("date")
)

server <- function(input, output, session) {

  observeEvent(input$b1, {

    add_notie_date_dialog(inputId = "date",
                          submitText = "Valider",
                          cancelText = "Annuler",
                          position = "top",
                          session)
  })

  output$date <- renderPrint(input$date)

}

# Run the application
shinyApp(ui = ui, server = server)
}
}
