% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notie-dependencies.R
\name{add_notie_deps}
\alias{add_notie_deps}
\title{notie dependencies utils}
\usage{
add_notie_deps()
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches notie. dependencies to the given tag
}
