% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_notie_input.R
\name{add_notie_input}
\alias{add_notie_input}
\title{Implemente la fonction date de la librairie JS notie}
\usage{
add_notie_input(
  inputId = "input",
  text = "Input notie",
  submitText = "Valider",
  cancelText = "Annuler",
  position = "top",
  autocapitalize = "none",
  autocomplete = "off",
  autocorrect = "off",
  autofocus = TRUE,
  inputmode = "latin",
  max = 10000,
  maxlength = NULL,
  min = NULL,
  minlength = NULL,
  placeholder = "Entrez un texte",
  value = "",
  spellcheck = FALSE,
  step = NULL,
  type = "text",
  session
)
}
\arguments{
\item{inputId}{id qui va contenir la reponse de l'appel à la notif (TRUE/FALSE)}

\item{text}{titre de l'input}

\item{submitText}{texte de la case de validation de la notif}

\item{cancelText}{texte de la case de refus de la notif}

\item{position}{position de la notif : string ('top', 'bottom')}

\item{autocapitalize}{boolean}

\item{autocomplete}{boolean}

\item{autocorrect}{boolean}

\item{autofocus}{boolean}

\item{inputmode}{char}

\item{max}{integer}

\item{maxlength}{integer}

\item{min}{integer}

\item{minlength}{integer}

\item{placeholder}{placeholder}

\item{value}{valeur par défaut}

\item{spellcheck}{boolean}

\item{step}{integer}

\item{type}{text}

\item{session}{argument session shiny}
}
\value{
une alerte notie
}
\description{
Implemente la fonction date de la librairie JS notie
}
\details{
https://github.com/jaredreich/notie

https://codingshiksha.com/javascript/notie-js-javascript-alert-toast-notification-library-full-tutorial-with-examples/
}
\examples{
\dontrun{library(shiny)
library(shinyYM)

ui <- fluidPage(
  add_notie_deps(),
  actionButton(inputId = "b1", "input"),
  verbatimTextOutput("input")
)

server <- function(input, output, session) {

  observeEvent(input$b1, {

    add_notie_input(inputId = "input",
                    text = "Input notie",
                    submitText = "Valider",
                    cancelText = "Annuler",
                    position = "top",
                    autocapitalize = "none",
                    autocomplete = "off",
                    autocorrect = "off",
                    autofocus = TRUE,
                    inputmode = "latin",
                    max = 10000,
                    maxlength = NULL,
                    min = NULL,
                    minlength = NULL,
                    placeholder = "Entrez un texte",
                    value = "",
                    spellcheck = FALSE,
                    step = NULL,
                    type = "text",
                    session)

  })

  output$input <- renderPrint(input$input)
}

# Run the application
shinyApp(ui = ui, server = server)
}
}
