% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater-jquery-dependencies.R
\name{add_rater_jquery_deps}
\alias{add_rater_jquery_deps}
\title{rater-jquery dependencies utils}
\usage{
add_rater_jquery_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches rater-jquery. dependencies to the given tag
}
