% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserui.R
\name{browserui}
\alias{browserui}
\title{Cree un bouton browser si on est en mode dev sur golem}
\usage{
browserui(inputId = NULL, ...)
}
\arguments{
\item{inputId}{inputId a utiliser}

\item{...}{parametres supplementaires pour l'actionButton}
}
\value{
un bouton Pause
}
\description{
Cree un bouton browser si on est en mode dev sur golem
}
\examples{
\dontrun{
browserui("toto")
}
}
