% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeWaiter.R
\name{closeWaiter}
\alias{closeWaiter}
\title{Fermeture de l'ecran de chargement avec logo}
\usage{
closeWaiter(isinProd, display_duration = 5)
}
\arguments{
\item{isinProd}{boolean indique si l'appli est en prod ou non (TRUE requis)}

\item{display_duration}{duree d'affichage de l'ecran de loading}
}
\value{
rien
}
\description{
Fermeture de l'ecran de chargement avec logo
}
