% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeWaiter_logoDatalab.R
\name{closeWaiter_logoDatalab}
\alias{closeWaiter_logoDatalab}
\title{Fermeture de l'ecran de chargement avec logo datalab}
\usage{
closeWaiter_logoDatalab(isinProd, display_duration = 5)
}
\arguments{
\item{isinProd}{boolean indique si l'appli est en prod ou non (TRUE requis)}

\item{display_duration}{duree d'affichage de l'ecran de loading}
}
\value{
rien
}
\description{
Fermeture de l'ecran de chargement avec logo datalab
}
