% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_div.R
\name{hidden_div}
\alias{hidden_div}
\title{Cree un div cache par defaut}
\usage{
hidden_div(id_div, contenu_div, ...)
}
\arguments{
\item{id_div}{id du div a cacher}

\item{contenu_div}{contenu du div a afficher ou masquer}

\item{...}{parametres supplementaires du div}
}
\value{
un taglist
}
\description{
Cree un div cache par defaut
}
\examples{
\dontrun{
hidden_div("toto", "toto")
}
}
