% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_some_ids.R
\name{hide_some_ids}
\alias{hide_some_ids}
\title{Permet de masquer une liste de div à partir de leurs ids}
\usage{
hide_some_ids(ids)
}
\arguments{
\item{ids}{une liste ou un vecteur}
}
\value{
rien
}
\description{
Permet de masquer une liste de div à partir de leurs ids
}
\examples{
\dontrun{hide_some_ids("coucou")}
}
