% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarPage_logoDatalab.R
\name{navbarPage_logoDatalab}
\alias{navbarPage_logoDatalab}
\title{Créer une navbarPage shiny avec le logo du datalab inséré}
\usage{
navbarPage_logoDatalab(
  title,
  style = "position:absolute;left:0\%; top:20\%;",
  ...
)
}
\arguments{
\item{title}{titre accole au logo}

\item{style}{css}

\item{...}{parametre supp}
}
\value{
un taglist
}
\description{
Créer une navbarPage shiny avec le logo du datalab inséré
}
\examples{
\dontrun{
# NE PAS UTILISER : pb avec les onglets obtenus
navbarPage_logoDatalab(title = "coucou")
}
}
