% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rating_input.R
\name{raterInput}
\alias{raterInput}
\title{Widget Rater pour Shiny}
\usage{
raterInput(
  inputId,
  label,
  initial_value,
  step_size,
  max_value,
  width = NULL,
  style = ""
)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label}

\item{initial_value}{valeur initiale}

\item{step_size}{pas entre 2 valeurs}

\item{max_value}{valeur maximale}

\item{width}{largeur du widget}

\item{style}{style css}
}
\value{
un widget shiny
}
\description{
Widget Rater pour Shiny
}
\examples{
\dontrun{
library(shiny)
library(shinyYM)

ui <- fluidPage(

  sidebarLayout(
    sidebarPanel(
      actionButton(inputId = "addone", "Ajouter 1 étoile"),
      actionButton(inputId = "substractone", "Retirer 1 étoile"),

      verbatimTextOutput("resultat1"),
      verbatimTextOutput("resultat2")

    ),

    mainPanel(
      raterInput("exemple1", label = "exemple1", initial_value = 1,
      step_size = .5, max_value = 10),
      raterInput("exemple2", label = "exemple2", initial_value = 4,
      step_size = 1, max_value = 8, style = "font-size: 35px; color : gold;")

    )
  )
)

server <- function(input, output, session) {

  output$resultat1 <- renderPrint(input$exemple1)
  output$resultat2 <- renderPrint(input$exemple2)

  observeEvent(input$addone, {
    updateRaterInput(inputId = "exemple1",
    value = pmin(10, input$exemple1 + 1), session = session)
    updateRaterInput(inputId = "exemple2",
    value = pmin(8, input$exemple2 + 1), session = session)
  })

  observeEvent(input$substractone, {
    updateRaterInput(inputId = "exemple1",
    value = pmax(0, input$exemple1 - 1), session = session)
    updateRaterInput(inputId = "exemple2",
    value = pmax(0, input$exemple2 - 1), session = session)
  })
}

shinyApp(ui = ui, server = server)
}
}
\references{
https://github.com/auxiliary/rater/
}
