% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_some_ids.R
\name{show_some_ids}
\alias{show_some_ids}
\title{Permet d'afficher une liste de div à partir de leurs ids}
\usage{
show_some_ids(ids)
}
\arguments{
\item{ids}{une liste ou un vecteur}
}
\value{
rien
}
\description{
Permet d'afficher une liste de div à partir de leurs ids
}
\examples{
\dontrun{show_some_ids("coucou")}
}
