% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_with_logoDatalab.R
\name{title_with_logoDatalab}
\alias{title_with_logoDatalab}
\title{Permet de créer un titre avec le logo du datalab acolle (par ex dans un navbarpage)}
\usage{
title_with_logoDatalab(main_title, style = "margin-top:-10\%;")
}
\arguments{
\item{main_title}{title colle au logo}

\item{style}{css eventuel}
}
\value{
un div
}
\description{
Permet de créer un titre avec le logo du datalab acolle (par ex dans un navbarpage)
}
\examples{
\dontrun{title_with_logoDatalab(main_title = "coucou")}
}
