% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rating_input.R
\name{updateRaterInput}
\alias{updateRaterInput}
\title{Mise à jour de la valeur du widget Rater}
\usage{
updateRaterInput(session, inputId, value)
}
\arguments{
\item{session}{objet shiny}

\item{inputId}{id de l'input}

\item{value}{valeur à définir}
}
\value{
un widget Rater
}
\description{
Mise à jour de la valeur du widget Rater
}
\examples{
\dontrun{
updateRaterInput(session, inputId = "rater", value = 12)
}
}
