% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiter_logo.R
\name{waiter_logo}
\alias{waiter_logo}
\title{Ajoute un ecran d'accueil shiny avec un logo}
\usage{
waiter_logo(isinProd, img_path)
}
\arguments{
\item{isinProd}{boolean indique si l'appli est en prod ou non (TRUE requis)}

\item{img_path}{lien vers l'image à utiliser}
}
\value{
rien
}
\description{
Ajoute un ecran d'accueil shiny avec un logo
}
\examples{
\dontrun{
waiter_logoDatalab(isinProd = TRUE)
}
}
