% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiter_logoDatalab.R
\name{waiter_logoDatalab}
\alias{waiter_logoDatalab}
\title{Ajoute un ecran d'accueil shiny avec le logo du Datalab (shiny : cote ui)}
\usage{
waiter_logoDatalab(isinProd)
}
\arguments{
\item{isinProd}{boolean indique si l'appli est en prod ou non (TRUE requis)}
}
\value{
rien
}
\description{
Ajoute un ecran d'accueil shiny avec le logo du Datalab (shiny : cote ui)
}
\examples{
\dontrun{
waiter_logoDatalab(isinProd = TRUE)
}
}
